pyproject.toml
[project]
name = "multi-llm-chat"
version = "0.1.0"
description = "AI Battle orchestration - adversarial multi LLM conversations with autoadaptive prompting"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.3",
    "aiosignal==1.4.0",
    "altair==5.5.0",
    "annotated-types==0.7.0",
    "anthropic==0.78.0",
    "anyio==4.12.1",
    "attrs==24.3.0",
    "beautifulsoup4==4.12.3",
    "black==25.1.0",
    "blinker==1.9.0",
    "cachetools==5.5.0",
    "certifi==2026.1.4",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "colorama==0.4.6 ; sys_platform == 'win32'",
    "coverage==7.7.1",
    "curated-tokenizers==0.0.9",
    "curated-transformers==0.1.1",
    "distro==1.9.0",
    "filelock==3.20.3",
    "flash-attention==1.0.0",
    "frozenlist==1.5.0",
    "fsspec==2026.2.0",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "google==3.0.0",
    "google-auth==2.37.0",
    "google-genai==1.41.0",
    "h11==0.16.0",
    "httpcore>=1.0.9",
    "httpx>=0.28.1",
    "huggingface-hub>=0.30.0",
    "idna==3.11",
    "iniconfig==2.1.0",
    "jinja2==3.1.6",
    "jiter==0.13.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2024.10.1",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "mlx>=0.25.2",
    "mlx-lm>=0.21.1",
    "mpmath==1.3.0",
    "multidict==6.1.0",
    "mypy-extensions==1.0.0",
    "narwhals==1.23.0",
    "networkx==3.6.1",
    "nlp==0.4.0",
    "numpy==2.2.2",
    "ollama>=0.6.0",
    "openai==2.17.0",
    "opencv-python==4.12.0.88",
    "packaging==24.2",
    "pandas==2.2.3",
    "pathspec==0.12.1",
    "phidata==2.7.10",
    "pillow==11.1.0",
    "plotly==6.0.1",
    "pluggy==1.5.0",
    "propcache==0.2.1",
    "psutil==7.0.0",
    "pyarrow==19.0.0",
    "pyasn1==0.6.2",
    "pyasn1-modules==0.4.1",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.8.1",
    "pydeck==0.9.1",
    "pygments==2.19.1",
    "pylint==3.3.6",
    "pytest==8.0.0",
    "pytest-cov==4.1.0",
    "python-dateutil==2.9.0.post0",
    "pytz==2024.2",
    "pyyaml==6.0.2",
    "referencing==0.36.1",
    "regex==2024.11.6",
    "requests==2.32.4",
    "rich==13.9.4",
    "rpds-py==0.22.3",
    "rsa==4.9",
    "safetensors==0.5.2",
    "scikit-learn==1.7.2",
    "scipy==1.15.2",
    "sentencepiece==0.2.1",
    "setuptools>=78.1.1",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "spacy==3.8.4",
    "spacy-curated-transformers==0.3.0",
    "spacytextblob==5.0.0",
    "streamlit==1.41.1",
    "sympy>=1.13.3",
    "textblob==0.19.0",
    "thinc==8.3.4",
    "tokenizers==0.21.0",
    "toml==0.10.2",
    "torch==2.9.0",
    "tqdm==4.67.1",
    "transformers==4.53.0",
    "typing-extensions==4.15.0",
    "tzdata==2025.1",
    "urllib3>=2.6.3",
    "watchdog==6.0.0",
    "watchdog==6.0.0 ; sys_platform != 'darwin'",
    "websockets==14.2",
    "yarl==1.18.3",
    "astroid==3.3.9",
    "blis==1.2.0",
    "catalogue==2.0.10",
    "cloudpathlib==0.21.0",
    "confection==0.1.5",
    "cymem==2.0.11",
    "dill==0.3.9",
    "docstring-parser==0.17.0",
    "isort==6.0.1",
    "joblib==1.4.2",
    "langcodes==3.5.0",
    "language-data==1.3.0",
    "marisa-trie==1.2.1",
    "mccabe==0.7.0",
    "murmurhash==1.0.12",
    "nltk==3.9.1",
    "platformdirs==4.3.7",
    "preshed==3.0.9",
    "python-dotenv==1.0.1",
    "shellingham==1.5.4",
    "smart-open==7.1.0",
    "spacy-legacy==3.0.12",
    "spacy-loggers==1.0.5",
    "srsly==2.5.1",
    "threadpoolctl==3.6.0",
    "tomli==2.2.1",
    "tomlkit==0.13.2",
    "typer==0.15.2",
    "wasabi==1.1.3",
    "weasel==0.4.1",
    "wrapt==1.17.2",
    "xxhash==3.5.0",
    "docker==7.1.0",
    "ffmpeg-python==0.2.0",
    "ffmpeg>=1.4",
    "mlx-metal>=0.29.4",
    "pycompile>=0.2.3",
]

[tool.uv.workspace]
members = ["mlx", "mlx-model"]

[tool.pylint.main]
extension-pkg-allow-list = ["cv2", "PIL"]
fail-under = 7.0

[tool.pylint."messages control"]
disable = [
    "logging-fstring-interpolation",
    "too-many-nested-blocks",
    "line-too-long",
    "broad-exception-caught",
    "too-many-instance-attributes",
    "too-many-arguments",
    "too-many-locals",
    "too-many-branches",
    "too-many-statements",
    "too-many-return-statements",
    "too-many-public-methods",
    "too-few-public-methods",
    "duplicate-code",
    "import-error",
    "no-name-in-module",
]

[tool.pylint.typecheck]
generated-members = ["cv2.*"]

[tool.pytest.ini_options]
pythonpath = ["src"]
